+++ changelog.md
@@ -1,3 +1,1 @@
-* Buying attacks has become more correct.
-* Events are now tracked correctly.
-* Minor changes.
+* Fix locale not items/currency.